requirements.txt
flask==3.0.0
openai==1.12.0
python-dotenv==1.0.0
edge-tts==6.1.12
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
werkzeug==3.0.1
pymysql==1.1.0

# Security packages
flask-wtf==1.2.1
flask-limiter==3.5.0
flask-talisman==1.1.0
flask-cors==4.0.0
bleach==6.1.0
